+++ src/interfaces/mining.h
@@ -8,6 +8,7 @@
 #include <consensus/amount.h>
 #include <interfaces/types.h>
 #include <node/types.h>
+#include <node/mining_types.h>
 #include <primitives/block.h>
 #include <primitives/transaction.h>
 #include <uint256.h>
@@ -21,9 +22,6 @@
 namespace node {
 struct NodeContext;
 } // namespace node
-
-class BlockValidationState;
-class CScript;
 
 namespace interfaces {
